+++ src/main/java/com/example/choyxona/repository/ProductKgClientRepository.java
@@ -20,9 +20,9 @@ public interface ProductKgClientRepository extends JpaRepository<ProductKgClient
 
     @Query(value = "select pr.name , sum(pc.kg) as amount , sum(pc.sum) as summa, sum(pc.cost) as costs from day d, client cl, product_kg pr, product_kg_client pc " +
             "where d.month_id = :monthId and d.id = cl.day_id and pr.id = pc.product_kg_id and pc.client_id = cl.id group by pr.name", nativeQuery = true)
-    List<ResPro_Client> getPrClientInfoMonth(int monthId);
+    List<ResPro_Client> getPrClientInfoMonth(long monthId);
 
     @Query(value = "select pr.name , sum(pc.kg) as amount , sum(pc.sum) as summa, sum(pc.cost) as costs from day d, client cl, product_kg pr, product_kg_client pc " +
             "where d.id = :dayId and d.id = cl.day_id and pr.id = pc.product_kg_id and pc.client_id = cl.id group by pr.name", nativeQuery = true)
-    List<ResPro_Client> getPrClientKgInfoDay(int dayId);
+    List<ResPro_Client> getPrClientKgInfoDay(long dayId);
 }
